requirements.txt
# Core dependencies
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0

# LangChain and AI
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# Streamlit UI
streamlit>=1.31.0

# Optional but recommended
sqlalchemy>=2.0.0